remove level of indirection when between boardview and its model also move related stuff into its own package
