+++ project/src/androidTest/java/de/earthlingz/oerszebra/InvalidmoveTest.java
@@ -6,6 +6,7 @@ import android.content.Intent;
 import android.test.ActivityInstrumentationTestCase2;
 
 import com.shurik.droidzebra.ZebraEngine;
+import de.earthlingz.oerszebra.BoardView.BoardViewModel;
 
 import java.lang.ref.WeakReference;
