+++ project/src/main/res/layout/board_layout.xml
@@ -5,7 +5,7 @@
   android:layout_width="fill_parent"
   android:layout_height="fill_parent">
 
-	<de.earthlingz.oerszebra.BoardView
+	<de.earthlingz.oerszebra.BoardView.BoardView
             android:id="@+id/board"
             android:layout_width="fill_parent"
             android:layout_height="wrap_content"
